refactor(controller): integrate sub-controllers into circuitEditorPageController/index.ts
